requirements.txt
absl-py==2.0.0
altair==5.1.2
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
attrs==23.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///C:/b/abs_0agyz1wsr4/croot/beautifulsoup4-split_1681493048687/work
blinker==1.6.2
brotlipy==0.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi @ file:///C:/ci_311/cffi_1676423759166/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1696431134479/work
click==8.1.7
colorama @ file:///C:/ci_311/colorama_1676422310965/work
comm @ file:///C:/ci_311/comm_1678376562840/work
cryptography @ file:///C:/b/abs_f4do8t8jfs/croot/cryptography_1694444424531/work
debugpy @ file:///C:/b/abs_c0y1fjipt2/croot/debugpy_1690906864587/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock @ file:///C:/ci_311/filelock_1676427284139/work
flatbuffers==23.5.26
gast==0.5.4
gdown @ file:///home/conda/feedstock_root/build_artifacts/gdown_1679726698720/work
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.9.0
idna @ file:///C:/ci_311/idna_1676424932545/work
importlib-metadata==6.8.0
jedi @ file:///C:/ci_311/jedi_1679427407646/work
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client @ file:///C:/b/abs_059idvdagk/croot/jupyter_client_1680171872444/work
jupyter_core @ file:///C:/b/abs_9d0ttho3bs/croot/jupyter_core_1679906581955/work
keras==2.14.0
libclang==16.0.6
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline @ file:///C:/ci_311/matplotlib-inline_1676425798036/work
mdurl==0.1.2
mistune==3.0.1
ml-dtypes==0.2.0
nbconvert==7.7.3
nest-asyncio @ file:///C:/ci_311/nest-asyncio_1676423519896/work
numpy==1.26.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging @ file:///C:/b/abs_28t5mcoltc/croot/packaging_1693575224052/work
pandas==2.1.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==10.0.1
platformdirs @ file:///C:/b/abs_b6z_yqw_ii/croot/platformdirs_1692205479426/work
prompt-toolkit @ file:///C:/ci_311/prompt-toolkit_1676425940920/work
protobuf==4.24.4
psutil @ file:///C:/ci_311_rebuilds/psutil_1679005906571/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.8.1b0
Pygments @ file:///C:/b/abs_fay9dpq4n_/croot/pygments_1684279990574/work
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2023.3.post1
pywin32==305.1
pyzmq @ file:///C:/b/abs_655zk4a3s8/croot/pyzmq_1686601465034/work
referencing==0.30.2
requests @ file:///C:/b/abs_316c2inijk/croot/requests_1690400295842/work
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.4
rsa==4.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
soupsieve @ file:///C:/b/abs_bbsvy9t4pl/croot/soupsieve_1696347611357/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
streamlit==1.27.2
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
toml==0.10.2
toolz==0.12.0
tornado @ file:///C:/b/abs_61jhmrrua1/croot/tornado_1690848767317/work
tqdm @ file:///C:/b/abs_f76j9hg7pv/croot/tqdm_1679561871187/work
traitlets @ file:///C:/ci_311/traitlets_1676423290727/work
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3 @ file:///C:/b/abs_889_loyqv4/croot/urllib3_1686163174463/work
validators==0.22.0
watchdog==3.0.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==3.0.0
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
wrapt==1.14.1
zipp==3.17.0
